+++ Dependiject/Resolver.swift
@@ -12,14 +12,19 @@ public protocol Resolver {
     /// Get a dependency of the specified type, using the name for disambiguation.
     /// - Precondition: A registration with the given type and name exists.
     /// - Returns: The instance of the type with the given name.
+    /// - Warning: This method is not thread-safe. In Swift 5.5+ (Xcode 13.2+), you may use
+    /// ``Resolver/resolve(_:name:)-4w8d6`` to ensure thread safety.
     func resolve<T>(_ type: T.Type, name: String?) -> T
 }
 
 public extension Resolver {
     /// Get an unnamed dependency of the specified type.
     /// - Returns: The unnamed instance of the type.
-    /// - Important: There must be an unnamed registration of the specified type. When resolving a
-    /// named dependency, use ``resolve(_:name:)`` or ``resolve(name:)`` instead.
+    ///
+    /// There must be an unnamed registration of the specified type. When resolving a named
+    /// dependency, use ``Resolver/resolve(_:name:)-7dkag`` or ``resolve(name:)`` instead.
+    /// - Warning: This method is not thread-safe. In Swift 5.5+ (Xcode 13.2+), you may use
+    /// ``Resolver/resolve(_:name:)-4w8d6`` to ensure thread safety.
     func resolve<T>(_ type: T.Type) -> T {
         return self.resolve(type, name: nil)
     }
@@ -29,8 +34,23 @@ public extension Resolver {
     /// - Returns: The instance of the type with the given name.
     /// 
     /// If the inferred type is incorrect, or if Swift cannot infer the type, use
-    /// ``resolve(_:name:)`` or ``resolve(_:)`` instead (which take the type as a parameter).
+    /// ``Resolver/resolve(_:name:)-7dkag`` or ``resolve(_:)`` instead (which take the type as a
+    /// parameter).
+    /// - Warning: This method is not thread-safe. In Swift 5.5+ (Xcode 13.2+), you may use
+    /// ``Resolver/resolve(_:name:)-4w8d6`` to ensure thread safety.
     func resolve<T>(name: String? = nil) -> T {
         return self.resolve(T.self, name: name)
     }
+    
+#if swift(>=5.5)
+    /// Get a dependency of the specified type, using the name for disambiguation.
+    /// - Precondition: A registration with the given type and name exists.
+    /// - Returns: The instance of the type with the given name.
+    @available(swift 5.5)
+    func resolve<T: Sendable>(_ type: T.Type = T.self, name: String? = nil) async -> T {
+        return await MainActor.run {
+            self.resolve(type, name: name)
+        }
+    }
+#endif
 }